RAG (Copy)/requirements.txt
pandas 
qdrant-client 
sentence-transformers 
tqdm
python-dotenv
groq
fastapi
uvicorn
streamlit